fix(mobile): handle magic-link deep link via an Expo Router file route
The Linking.useURL hook raced Expo Router, which routed the deep link to a
non-existent /verify route and showed "Unmatched Route". Replace it with an
app/auth/verify/[token] file route that Expo Router maps from both the App Link
(https://<host>/auth/verify/<token>) and the custom scheme
(biketripplanner://auth/verify/<token>).
