+++ mobile/app/_layout.tsx
@@ -1,16 +1,15 @@
 import { Stack } from 'expo-router';
 import { StatusBar } from 'expo-status-bar';
 import { AuthProvider } from '../src/auth/store';
-import { useDeepLinkAuth } from '../src/auth/useDeepLinkAuth';
 
 function RootNavigator() {
-  useDeepLinkAuth();
   return (
     <Stack>
       <Stack.Screen name="index" options={{ headerShown: false }} />
       <Stack.Screen name="login" options={{ title: 'Connexion' }} />
       <Stack.Screen name="(tabs)" options={{ headerShown: false }} />
       <Stack.Screen name="trip/[id]" options={{ title: 'Roadbook' }} />
+      <Stack.Screen name="auth/verify/[token]" options={{ headerShown: false }} />
     </Stack>
   );
 }
